add extra instruction to live streams

+++ src/musicplayer/MusicPlayer.Controller.ts
@@ -13,10 +13,11 @@ export const commandController = async (ctx: MusicPlayer, message: Message) => {
     if (delegation) {
         clearTimeout(ctx.idleTimer);
         delegation();
-        setTimeout(
-            () => message.delete().catch((noop) => noop),
-            configObj.autoDeleteAfter
-        );
+        if (configObj.autoDeleteAfter)
+            setTimeout(
+                () => message.delete().catch((noop) => noop),
+                configObj.autoDeleteAfter * 1000
+            );
     }
 };
